MAINT: fix compile errors.

+++ cpp/examples/Sara/ImageProcessing/Chessboard/EdgeFusion.cpp
@@ -109,7 +109,7 @@ namespace DO::Sara {
       const auto frame_blurred = frame.convert<Rgb32f>()
                                       .compute<Gaussian>(1.2f)
                                       .convert<Rgb8>();
-      const auto color_threshold = std::sqrt(std::powf(1, 2) * 3);
+      const auto color_threshold = std::sqrt(std::pow(1, 2) * 3);
       const auto regions = color_watershed(frame_blurred, color_threshold);
 
       // Display the good regions.
